Freshen up README
Hi @AaronHemphill @AlphaTiles/alphatilesdevelopers !
Here are some "take it or leave it" suggestions for the README. I noticed that the section on RTL support was outdated (I forgot to update it when I did away with the "mirror-flip" integer setting). I wanted to replace some of the urls with in-text hyperlinks. And I thought some things could be phrased differently, but that's not super important. Best -Lainey